BIG COMMIT.  - add classes Vertex VertexList VertKDT.  - OBJ file format(ascii and binary) is ready to use.  - Vertex data is moved from inside PrivateTriangle to Vertex.  - VertexList Holds Vertex instance pointers.  - Triangle class holds the pointers to Vertex.  - We have a function of removing same vertex in VertexList.  - At removing save vertex, if vertex is same or not checked    by using KDTree(VertKDT).  - MPIcodes are changed appropiate.  - MPI Type Problems are managed by sizeof(T).

+++ include/common/Vec3.h
@@ -121,7 +121,8 @@ public:
 	static Vec3<T> zaxis() { return Vec3<T>(0, 0, 1); }
 
 	T lengthSquared() const { return t[0] * t[0] + t[1] * t[1] + t[2] *t [2]; }
-	T length() const { return sqrtf(lengthSquared()); }
+	//T length() const { return sqrtf(lengthSquared()); }
+	T length() const { return sqrt(lengthSquared()); }
 	Vec3<T>& normalize() {
 		T len = length();
 		if (len != 0)
@@ -136,7 +137,8 @@ public:
 		else
 			return *this;
 	}
-	T average() const { return (t[0] + t[1] + t[2])/3.f; }
+	//T average() const { return (t[0] + t[1] + t[2])/3.f; }
+	T average() const { return (t[0] + t[1] + t[2])/3.; }
 
 	T t[3];
 };
